fix bug: correct rounding p-s with neg tests
the error test didn't take into account negative test statistics when determining the upper and lower bound of the p-values. when the test statistic was negative, these bounds were switched, which resulted in incorrectly flagged errors.

the upper and lower bound of the p-value are now classified based on which of the two p-values is the lowest.