simplify regex for chi2
This regex contained several unnecessary statements. It's simplified to a point where it simply extracts something that looks like a statistical test, but isn't one of a set of other specified letters/tests. Updating this regex did not negatively affect the unit tests.

It did turn out that this simplification finds 1 extra chi-2 (a delta-G) in Nuijten et al., so this had to be added to the unit tests.