don't read parse "correlations" > 1
in 1 paper (aloe.pdf) this caused an error, because QError = 26.06 ... was read as a correlation. This is the first half of the fix to avoid this: don't include correlations > 1. The second part of the fix (not yet implemented) is to not include subscripts ending on r as a correlation (same holds for other tests, i.e., subscript ending on t should not be read as t test)