only extract stats if no letters precede them
adapted the regexes of the test statistics so that they are only recognized if they are not preceded by a letter. This is necessary to avoid things like "QError(14).." being recognized as 'r(14)'. This also solved the weird hack to extract z stats but only if there was a space in front of them. Furthermore, this regex does still match stats that are between parentheses/quotes etc.